refactor: move ffi api to own file, not init.py
